requirements.txt
# requirements.txt
snscrape
python-dotenv
